fix: rename shared mod to core
